wip refactoring / cleanup
